requirements.txt
Django==3.1.7
psycopg2-binary>=2.8
django-taggit==1.3.0
python-ldap==3.1.0
django-mathfilters
requests==2.25.1
beautifulsoup4==4.9.3
lxml==4.6.3
django-bootstrap-modal-forms
XlsxWriter
gunicorn==20.1.0
